Button synchronization fixed

+++ js/main.js
@@ -6,25 +6,34 @@ $(document).ready(function () {
 
     var btnStart = $('#btnTimerStart');
     var btnReset = $('#btnTimerReset');
-   /* var hours = $('#hours');*/
+    /* var hours = $('#hours');*/
     var minutes = $('#minutes');
     var seconds = $('#seconds');
     var milli = $('#milliSeconds');
+
+    //lap one buttons
     var lapOneBtnStart = $('#lapOneStart');
     var lapOneBtnStop = $('#lapOneStop');
-   /* var lapOneHour = $('#lapOneHour');*/
+    var lapOneBtnHumanIntervention = $('#lapOneBtnHandIntervention');
+    /* var lapOneHour = $('#lapOneHour');*/
     var lapOneMinute = $('#lapOneMinute');
     var lapOneSecond = $('#lapOneSecond');
     var lapOneMilliSecond = $('#lapOneMilliSecond');
+
+
+    //lap two buttons
     var lapTwoBtnStart = $('#lapTwoBtnStart');
     var lapTwoBtnStop = $('#lapTwoBtnStop');
-   /* var lapTwoHour = $('#lapTwoHour');*/
+    var lapTwoBtnHumarIntervention = $('#lapTwoBtnHandIntervention');
+    /* var lapTwoHour = $('#lapTwoHour');*/
     var lapTwoMinute = $('#lapTwoMinute');
     var lapTwoSecond = $('#lapTwoSecond');
     var lapTwoMilliSecond = $('#lapTwoMilliSecond');
+
+    //lap three buttons
     var lapThreeBtnStart = $('#lapThreeBtnStart');
     var lapThreeBtnStop = $('#lapThreeBtnStop');
-   /* var lapThreeHour = $('#lapThreeHour');*/
+    /* var lapThreeHour = $('#lapThreeHour');*/
     var lapThreeMinute = $('#lapThreeMinute');
     var lapThreeSecond = $('#lapThreeSecond');
     var lapThreeMilliSecond = $('#lapThreeMilliSecond');
@@ -41,23 +50,35 @@ $(document).ready(function () {
     var lapTwoWatch = new StopWatchTimer(lapTwoMilliSecond, lapTwoSecond, lapTwoMinute, false);
     var lapThreeWatch = new StopWatchTimer(lapThreeMilliSecond, lapThreeSecond, lapThreeMinute, false);
 
+    var lapOneFlag = false;
+    var lapTwoFlag = false;
+    var lapThreeFlag = false;
+    var mainFlagInterrupt = 0;
+    var countHumanIntervention = 0;
+
+
     function setTrashBtnVisibility() {
         if (localStorage.length > 0) {
             btnTrash.attr('class', 'btn btn-danger active');
             downLoadAnchor.attr("class", "btn btn-primary active");
+            trashFlag = true;
         }
     }
 
     setTrashBtnVisibility();
 
+
+    //trash button click
     btnTrash.click(function () {
-        $('#delete-file-modal').modal();
-        $('#confirm-delete-button').click(function () {
-            localStorage.clear();
-            btnTrash.attr('class', 'btn btn-danger disabled');
-            downLoadAnchor.attr("class", "btn btn-primary disabled");
-            showModel("Success", "Previous all stored data has been deleted");
-        });
+        if (trashFlag) {
+            $('#delete-file-modal').modal();
+            $('#confirm-delete-button').click(function () {
+                localStorage.clear();
+                btnTrash.attr('class', 'btn btn-danger disabled');
+                downLoadAnchor.attr("class", "btn btn-primary disabled");
+                showModel("Success", "Previous all stored data has been deleted");
+            });
+        }
     });
 
     btnStart.click(function () {
@@ -67,6 +88,7 @@ $(document).ready(function () {
             mainWatch.start();
             btnText = 'Stop';
             btnAttr = 'btn btn-danger active';
+            mainFlagInterrupt = 1;
             disableReset();
 
         }
@@ -74,6 +96,7 @@ $(document).ready(function () {
             mainWatch.stop();
             btnText = 'Start';
             btnAttr = 'btn btn-success active';
+            mainFlagInterrupt = 2;
             enableReset();
         }
         btnStart.text(btnText);
@@ -82,17 +105,15 @@ $(document).ready(function () {
     });
 
     btnReset.click(function () {
-        teamName.val("");
-        mainWatch.reset();
-        lapOneWatch.reset();
-        lapTwoWatch.reset();
-        lapThreeWatch.reset();
-        btnSave.attr('class', 'btn btn-primary disabled');
-        btnCancel.attr('class', 'btn btn-danger disabled');
+        if (mainFlagInterrupt == 2) {
+            resetAll();
+        }
     });
 
     function disableReset() {
         btnReset.attr('class', 'btn btn-default disabled');
+        btnSave.attr('class', 'btn btn-primary disabled');
+        btnCancel.attr('class', 'btn btn-danger disabled');
         enableOtherLap();
         /*lapOneBtnStart.attr('class', 'btn btn-primary active');*/
     };
@@ -112,80 +133,122 @@ $(document).ready(function () {
     function enableReset() {
         btnReset.attr('class', 'btn btn-default active');
         disableOtherLap();
+        btnSave.attr('class', 'btn btn-primary active');
+        btnCancel.attr('class', 'btn btn-danger active');
 
         /*lapOneBtnStart.attr('class', 'btn btn-primary disabled');*/
     };
 
     function disableOtherLap() {
-        if (lapOneWatch.state == 1) {
+        if (lapOneWatch.state == 1 || lapOneWatch.state == 0) {
             lapOneWatch.stop();
             lapOneWatch.state = 3;
-            lapOneBtnStop.attr('class', 'btn btn-danger disabled');
         }
-        else if (lapTwoWatch.state == 1) {
+        else if (lapTwoWatch.state == 1 || lapTwoWatch.state == 0) {
             lapTwoWatch.stop();
             lapTwoWatch.state = 3;
-            lapTwoBtnStop.attr('class', 'btn btn-danger disabled');
         }
-        else if (lapTwoWatch.state == 0) {
-            lapTwoBtnStart.attr('class', 'btn btn-primary disabled');
-        }
-        else if (lapThreeWatch.state == 0) {
-            lapThreeBtnStart.attr('class', 'btn btn-primary disabled');
-        }
-        else if (lapThreeWatch.state == 1) {
+        else if (lapThreeWatch.state == 1 || lapThreeWatch.state == 0) {
             lapThreeWatch.stop();
             lapThreeWatch.state = 3;
-            lapThreeBtnStop.attr('class', 'btn btn-danger disabled');
         }
-    }
-
-    lapOneBtnStart.click(function () {
-
-        lapOneWatch.start();
         lapOneBtnStart.attr('class', 'btn btn-primary disabled');
-        lapOneBtnStop.attr('class', 'btn btn-danger active');
+        lapOneBtnStop.attr('class', 'btn btn-danger disabled');
+        lapOneBtnHumanIntervention.attr('class', 'btn btn-warning disabled');
+        lapTwoBtnStart.attr('class', 'btn btn-primary disabled');
+        lapTwoBtnStop.attr('class', 'btn btn-danger disabled');
+        lapTwoBtnHumarIntervention.attr('class', 'btn btn-warning disabled');
+        lapThreeBtnStart.attr('class', 'btn btn-primary disabled');
+        lapThreeBtnStop.attr('class', 'btn btn-danger disabled');
+    };
 
+
+    //lapOne button onClick methods
+    lapOneBtnStart.click(function () {
+        if (mainFlagInterrupt == 1 && (lapOneWatch.state == 0 || lapOneWatch.state == 3)) {
+            lapOneWatch.start();
+            lapOneBtnStart.attr('class', 'btn btn-primary disabled');
+            lapOneBtnStop.attr('class', 'btn btn-danger active');
+            lapOneBtnHumanIntervention.attr('class', 'btn btn-warning active');
+        }
     });
 
     lapOneBtnStop.click(function () {
-        lapOneWatch.stop();
-        lapOneBtnStop.attr('class', 'btn btn-danger disabled');
-        lapTwoBtnStart.attr('class', 'btn btn-primary active');
+        if (lapOneWatch.state == 1) {
+            stopLapOne();
+        }
     });
 
+    lapOneBtnHumanIntervention.click(function () {
+        if (lapOneWatch.state == 1) {
+            countHumanIntervention++;
+            stopLapOne();
+        }
+    });
+
+    function stopLapOne() {
+        lapOneWatch.stop();
+        lapOneBtnStop.attr('class', 'btn btn-danger disabled');
+        lapOneBtnHumanIntervention.attr('class', 'btn btn-danger disabled');
+        lapTwoBtnStart.attr('class', 'btn btn-primary active');
+        lapOneFlag = true;
+    };
+
+
+    //lap Two button onClick methods
     lapTwoBtnStart.click(function () {
-
-        lapTwoWatch.start();
-        lapTwoBtnStart.attr('class', 'btn btn-primary disabled');
-        lapTwoBtnStop.attr('class', 'btn btn-danger active');
-
+        if (lapOneFlag && mainFlagInterrupt == 1 && (lapTwoWatch.state == 0 || lapTwoWatch.state == 3)) {
+            lapTwoWatch.start();
+            lapTwoBtnStart.attr('class', 'btn btn-primary disabled');
+            lapTwoBtnStop.attr('class', 'btn btn-danger active');
+            lapTwoBtnHumarIntervention.attr('class', 'btn btn-warning active');
+        }
     });
 
     lapTwoBtnStop.click(function () {
-        lapTwoWatch.stop();
-        lapTwoBtnStop.attr('class', 'btn btn-danger disabled');
-        lapThreeBtnStart.attr('class', 'btn btn-primary active');
+        if (lapTwoWatch.state == 1) {
+            stopLapTwo();
+        }
     });
 
+    lapTwoBtnHumarIntervention.click(function () {
+        if (lapTwoWatch.state == 1) {
+            countHumanIntervention++;
+            stopLapTwo();
+        }
+    });
+
+    function stopLapTwo() {
+        lapTwoWatch.stop();
+        lapTwoBtnStop.attr('class', 'btn btn-danger disabled');
+        lapTwoBtnHumarIntervention.attr('class', 'btn btn-warning disabled');
+        lapThreeBtnStart.attr('class', 'btn btn-primary active');
+        lapTwoFlag = true;
+    }
+
+
+    //lap three buttons onClick methods
     lapThreeBtnStart.click(function () {
-
-        lapThreeWatch.start();
-        lapThreeBtnStart.attr('class', 'btn btn-primary disabled');
-        lapThreeBtnStop.attr('class', 'btn btn-danger active');
-
+        if (lapTwoFlag && mainFlagInterrupt == 1 && (lapThreeWatch.state == 0 || lapThreeWatch.state == 3)) {
+            lapThreeWatch.start();
+            lapThreeBtnStart.attr('class', 'btn btn-primary disabled');
+            lapThreeBtnStop.attr('class', 'btn btn-danger active');
+        }
     });
 
     lapThreeBtnStop.click(function () {
-        lapThreeWatch.stop();
-        lapThreeBtnStop.attr('class', 'btn btn-danger disabled');
-        btnSave.attr('class', 'btn btn-primary active');
-        btnCancel.attr('class', 'btn btn-danger active');
+        if (lapThreeWatch.state == 1) {
+            lapThreeWatch.stop();
+            lapThreeBtnStop.attr('class', 'btn btn-danger disabled');
+            lapThreeFlag = true;
+        }
     });
 
+
+    // data saving buttons
     btnSave.click(function () {
         console.log("btn save click");
-        if (localStorage) {
+        if (localStorage && mainFlagInterrupt == 2) {
             var team = teamName.val();
             if (team != null && team != '') {
                 /*console.log("Team name found" + team);*/
@@ -211,6 +274,27 @@ $(document).ready(function () {
             }
         }
     });
+
+    btnCancel.click(function () {
+        if (mainFlagInterrupt == 2) {
+            resetAll();
+        }
+    });
+
+    //reset all
+    function resetAll() {
+        teamName.val("");
+        mainWatch.reset();
+        lapOneWatch.reset();
+        lapTwoWatch.reset();
+        lapThreeWatch.reset();
+        btnSave.attr('class', 'btn btn-primary disabled');
+        btnCancel.attr('class', 'btn btn-danger disabled');
+        countHumanIntervention = 0;
+        lapOneFlag = false;
+        lapTwoFlag = false;
+        lapThreeFlag = false;
+    };
 
     function showModel(modelType, message) {
         var modelHeader = $('.modal-title');
@@ -238,25 +322,6 @@ $(document).ready(function () {
         popUpDialog.modal();
 
     };
-
-    /* function setSessionDataEnabled() {
-     if (typeof(Storage) !== "undefined") {
-     if (!sessionStorage.enabled) {
-     sessionStorage.enabled = 1;
-     }
-     }
-     };*/
-
-
-    btnCancel.click(function () {
-        teamName.val("");
-        mainWatch.reset();
-        lapOneWatch.reset();
-        lapTwoWatch.reset();
-        lapThreeWatch.reset();
-        btnSave.attr('class', 'btn btn-primary disabled');
-        btnCancel.attr('class', 'btn btn-danger disabled');
-    });
 
 
     function setDownload() {
